+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/ColorSemanticTokens.swift
@@ -327,65 +327,6 @@ public protocol ColorSemanticTokens {
     var colorDecorativeNeutralMutedLowest: ColorSemanticToken { get }
     var colorDecorativeNeutralMutedMedium: ColorSemanticToken { get }
 
-    // MARK: - Color - Chart
-
-    var colorChartBorderLight: ColorSemanticToken { get }
-    var colorChartBorderDark: ColorSemanticToken { get }
-
-    var colorChartBorderContrastLight: ColorSemanticToken { get }
-    var colorChartBorderContrastDark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier1Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier1Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier2Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier2Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier3Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier3Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier4Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier4Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier5Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier5Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier6Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier6Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier7Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier7Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier8Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier8Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier9Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier9Dark: ColorSemanticToken { get }
-
-    var colorChartCategoricalTier10Light: ColorSemanticToken { get }
-    var colorChartCategoricalTier10Dark: ColorSemanticToken { get }
-
-    var colorChartFunctionalInfoLight: ColorSemanticToken { get }
-    var colorChartFunctionalInfoDark: ColorSemanticToken { get }
-
-    var colorChartFunctionalNegativeLight: ColorSemanticToken { get }
-    var colorChartFunctionalNegativeDark: ColorSemanticToken { get }
-
-    var colorChartFunctionalPositiveLight: ColorSemanticToken { get }
-    var colorChartFunctionalPositiveDark: ColorSemanticToken { get }
-
-    var colorChartFunctionalWarningLight: ColorSemanticToken { get }
-    var colorChartFunctionalWarningDark: ColorSemanticToken { get }
-
-    var colorChartGridlinesLight: ColorSemanticToken { get }
-    var colorChartGridlinesDark: ColorSemanticToken { get }
-
-    var colorChartHighlightLight: ColorSemanticToken { get }
-    var colorChartHighlightDark: ColorSemanticToken { get }
-
-    var colorChartNeutralLight: ColorSemanticToken { get }
-    var colorChartNeutralDark: ColorSemanticToken { get }
-
     // MARK: - Color - Repository
 
     var colorRepositoryAccentDefault: ColorSemanticToken { get }